refactor hoc and selectors and notify draft saved
settings and with notification
