+++ src/frontend/web_application/locales/en/main.json
@@ -112,6 +112,7 @@
     "device.verify.descr": "You need to verify this device to allow it to access to your Caliopen account. Please note that this action can have an impact on your Privacy Index according to the device information and settings. You'll find below some informations about the device which tried to connect to your Caliopen account.",
     "device.verify.not-you": "It's not you?",
     "device.verify.title": "Verify you device",
+    "draft.feedback.saved": "Draft saved",
     "general.action.cancel": "Cancel",
     "general.action.load_more": "Load more",
     "header.menu.account": "Account",
