+++ src/frontend/web_application/locales/fr/main.json
@@ -112,6 +112,7 @@
     "device.verify.descr": "Vous devez vérifier cet appareil afin de lui permettre de se connecter à votre compte Caliopen. Merci de noter que cette opération a un impact sur l'Index de Confidentialité selon ses informations et paramétrages. Vous trouverez ci-dessous quelques informations à propos de cet appareil, qui a essayé de se connecter à votre compte Caliopen.",
     "device.verify.not-you": "Ce n'est pas le vôtre ?",
     "device.verify.title": "Vérifier votre appareil",
+    "draft.feedback.saved": "Brouillon sauvegardé",
     "general.action.cancel": "Annuler",
     "general.action.load_more": "Charger plus",
     "header.menu.account": "Compte",
