+++ src/frontend/web_application/locales/template/main.json
@@ -108,6 +108,7 @@
     "device.verify.descr": "device.verify.descr",
     "device.verify.not-you": "device.verify.not-you",
     "device.verify.title": "device.verify.title",
+    "draft.feedback.saved": "draft.feedback.saved",
     "general.action.cancel": "general.action.cancel",
     "general.action.load_more": "general.action.load_more",
     "header.menu.account": "header.menu.account",
