+++ src/frontend/web_application/src/layouts/Page/components/CallToAction/components/ReplyButton/index.js
@@ -1,7 +1,7 @@
 import { compose } from 'redux';
 import { withTranslator } from '@gandi/react-translate';
 import Presenter from './presenter';
-import { withNotification } from '../../../../../../services/notification';
+import { withNotification } from '../../../../../../hoc/notification';
 
 export default compose(
   withNotification(),
