+++ src/frontend/web_application/src/scenes/ContactBook/components/ContactList/index.js
@@ -1,8 +1,8 @@
 import { createSelector } from 'reselect';
 import { connect } from 'react-redux';
 import Presenter from './presenter';
+import { settingsSelector } from '../../../../store/selectors/settings';
 
-const settingsSelector = state => state.settings.settings;
 const mapStateToProps = createSelector(
   [settingsSelector],
   ({ contact_display_order, contact_display_format }) => ({
