+++ src/frontend/web_application/src/scenes/ContactBook/components/ImportContact/index.js
@@ -2,7 +2,7 @@ import { bindActionCreators, compose } from 'redux';
 import { connect } from 'react-redux';
 import { withTranslator } from '@gandi/react-translate';
 import { requestContacts } from '../../../../store/modules/contact';
-import { withNotification } from '../../../../services/notification';
+import { withNotification } from '../../../../hoc/notification';
 import Presenter from './presenter';
 
 const mapDispatchToProps = dispatch => bindActionCreators({
