+++ src/frontend/web_application/src/store/configure-store.js
@@ -6,7 +6,7 @@ import axiosMiddleware from './middlewares/axios-middleware';
 import contactMiddleware from './middlewares/contacts-middleware';
 import deviceMiddleware from './middlewares/device-middleware';
 import discussionMiddleware from './middlewares/discussions-middleware';
-import draftMessageMiddleware from './middlewares/draft-messages-middleware';
+import draftMessageMiddleware, { middlewares as draftMessageMiddlewares } from './middlewares/draft-messages-middleware';
 import i18nMiddleware from './middlewares/i18n-middleware';
 import importanceLevelMiddleware from './middlewares/importance-level-middleware';
 import messageMiddleware from './middlewares/messages-middleware';
@@ -24,6 +24,7 @@ const middlewares = [
   deviceMiddleware,
   discussionMiddleware,
   draftMessageMiddleware,
+  ...draftMessageMiddlewares,
   i18nMiddleware,
   importanceLevelMiddleware,
   messageMiddleware,
