+++ src/frontend/web_application/src/store/selectors/settings.js
@@ -1,0 +1,1 @@
+export const settingsSelector = state => state.settings.settings;
